Phase 39 complete: paper queue ALL GREEN (43/43), NMF module, cleanup
Paper queue extension (Exp152-156): physical comm modes, nitrifying QS,
marine interkingdom QS, Myxococcus critical density, Dictyostelium relay.

Drug repurposing track (Exp157-161): Fajgenbaum pathway scoring, MATRIX
pharmacophenomics, NMF drug-disease factorization, repoDB reproduction,
knowledge graph embedding baseline.

New module: bio::nmf — Lee & Seung multiplicative updates (Euclidean + KL),
cosine similarity, top-K predictions. 4 unit tests. ToadStool absorption
target for GPU shader generation.

Cleanup: updated stale WGSL shader references in EVOLUTION_READINESS,
DEPRECATION_MIGRATION, BARRACUDA_REQUIREMENTS. Fixed V25 handoff queued
papers section. Moved drug_repurposing/ script to scripts/.

161 experiments, 3,132+ checks, 732 tests, all PASS.

+++ barracuda/src/bio/batch_fitness_gpu.rs
@@ -105,7 +105,7 @@ mod tests {
         let weights = vec![3.0f32, 5.0];
         let result = bf.evaluate(&pop, &weights, 2, 2).expect("fitness");
         assert_eq!(result.len(), 2);
-        assert!((result[0] - 3.0).abs() < tolerances::GPU_F32_PARITY); // 1*3 + 0*5
-        assert!((result[1] - 5.0).abs() < tolerances::GPU_F32_PARITY); // 0*3 + 1*5
+        assert!((result[0] - 3.0).abs() < tolerances::GPU_F32_PARITY as f32); // 1*3 + 0*5
+        assert!((result[1] - 5.0).abs() < tolerances::GPU_F32_PARITY as f32); // 0*3 + 1*5
     }
 }
